# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Profy Skill Academy API
  version: 0.1.0
  license:
    name: UNLICENSED
    identifier: UNLICENSED
  description: |
    REST API for the Profy Skill Academy learning platform (TECHNICAL_DOC §5).

    * Consumer endpoints (mobile + web share them) accept a JWT with audience
      `consumer`. Admin endpoints live under `/api/v1/admin/*` and accept a JWT
      with audience `admin`. The two token types are never interchangeable.
    * Errors use the envelope `{"error": {"code": string, "message": string}}`
      with stable machine codes (`ai_unavailable`, `rate_limited`, ...).
    * Clients send an `X-App-Version` header; the API can force-update later.
servers:
  - url: http://localhost:8080
    description: Local development (docker compose)
...
paths:
...
  /api/v1/courses/{slug}:
    get:
      tags: [lessons]
      summary: Course detail with lessons grouped by level
      operationId: courseDetail
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Course and its published lessons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/lessons/{slug}:
    get:
      tags: [lessons]
      summary: Full lesson (all AI-scope fields + primary video + alternates)
      operationId: lessonDetail
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lesson content; drafts are never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ------------------------------------------------------------------ AI Teacher
...
  /api/v1/lessons/{id}/progress:
    put:
      tags: [progress]
      summary: Update lesson progress (course completion is derived)
      operationId: updateProgress
      security:
        - consumerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressRequest"
      responses:
        "204":
          description: Saved
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/progress/continue:
    get:
      tags: [progress]
      summary: Continue-learning shortcut (most recent in-progress lessons)
      operationId: continueLearning
      security:
        - consumerAuth: []
      responses:
        "200":
          description: Recent lessons to resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinueLearningResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/library/bookmarks:
    get:
      tags: [progress]
      summary: List bookmarks
      operationId: listBookmarks
      security:
        - consumerAuth: []
      responses:
        "200":
          description: Bookmarks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/library/bookmarks/{lessonId}:
    post:
      tags: [progress]
      summary: Add a bookmark (idempotent)
      operationId: addBookmark
      security:
        - consumerAuth: []
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Bookmarked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [progress]
      summary: Remove a bookmark (idempotent)
      operationId: removeBookmark
      security:
        - consumerAuth: []
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Removed
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/lessons/{id}/quiz-attempts:
    post:
      tags: [progress]
      summary: Record a quiz attempt (scoring is client-side at MVP)
      operationId: recordQuizAttempt
      security:
        - consumerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAttemptRequest"
      responses:
        "201":
          description: Attempt recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/profile/stats:
    get:
      tags: [progress]
      summary: Learner stats for the profile screen
      operationId: profileStats
      security:
        - consumerAuth: []
      responses:
        "200":
          description: Aggregated stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileStats"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # -------------------------------------------------------------------- billing
...
  /api/v1/admin/lessons:
    get:
      tags: [admin]
      summary: List lessons (drafts included)
      operationId: adminListLessons
      security:
        - adminAuth: []
      parameters:
        - name: node_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminLesson"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [admin]
      summary: Create a lesson
      operationId: adminCreateLesson
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLesson"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/admin/lessons/{id}:
    get:
      tags: [admin]
      summary: Get a lesson with its videos
      operationId: adminGetLesson
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLesson"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [admin]
      summary: Update a lesson (all content fields incl. quizzes)
      operationId: adminUpdateLesson
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLesson"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [admin]
      summary: Delete a lesson
      operationId: adminDeleteLesson
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/lessons/{id}/videos:
    get:
      tags: [admin]
      summary: List video candidates for a lesson
      operationId: adminListVideos
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoCandidate"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [admin]
      summary: Attach a curated YouTube video (URL pasted → validated)
      operationId: adminAddVideo
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "201":
          description: Attached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoCandidate"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/admin/lessons/{id}/videos/{videoId}:
    put:
      tags: [admin]
      summary: Update video status, primary/alternate flags, review date, notes
      operationId: adminUpdateVideo
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: videoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoCandidate"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [admin]
      summary: Detach a video
      operationId: adminDeleteVideo
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: videoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Detached
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ------------------------------------------------------ review & audit (admin)
  /api/v1/admin/review/videos:
    get:
      tags: [admin]
      summary: Review queue — flagged/unavailable videos with last check result
      operationId: adminReviewVideos
      security:
        - adminAuth: []
      responses:
        "200":
          description: Queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewVideoItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    consumerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Consumer JWT with audience `consumer` (15 min).
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT with audience `admin` (15 min). Never interchangeable with consumer tokens.
  responses:
    Unauthorized:
      description: Missing/invalid/expired token, or token audience mismatch
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found (or invisible to this audience)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Invalid request body/parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine code, e.g. ai_unavailable, rate_limited.
            message:
              type: string
...
    CourseSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: [string, "null"]
        icon:
          type: [string, "null"]
        lesson_count:
          type: integer
        levels:
          type: array
          items:
            type: string
            enum: [beginner, intermediate, advanced]
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/CourseSummary"
        - type: object
          properties:
            lessons:
              type: array
              items:
                $ref: "#/components/schemas/LessonSummary"
    LessonSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        level:
          type: [string, "null"]
          enum: [beginner, intermediate, advanced]
        sort_order:
          type: integer
    LessonDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        description:
          type: [string, "null"]
        explanation:
          type: [string, "null"]
        objectives:
          type: array
          items:
            type: string
        examples:
          type: array
          items:
            type: string
        exercises:
          type: array
          items:
            type: string
        quizzes:
          type: array
          items:
            $ref: "#/components/schemas/Quiz"
        level:
          type: [string, "null"]
          enum: [beginner, intermediate, advanced]
        video:
          $ref: "#/components/schemas/LessonVideo"
        alternates:
          type: array
          items:
            $ref: "#/components/schemas/LessonVideo"
    LessonVideo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        youtube_video_id:
          type: string
          pattern: "^[A-Za-z0-9_-]{11}$"
        title:
          type: string
        channel:
          type: [string, "null"]
        is_primary:
          type: boolean
    Quiz:
      type: object
      required: [question, options, answer_index]
      properties:
        question:
          type: string
        options:
          type: array
          minItems: 2
          items:
            type: string
        answer_index:
          type: integer
          minimum: 0

    # ------------------------------------------------------------ discovery
...
    ProgressRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [in_progress, completed]
    ContinueLearningItem:
      type: object
      properties:
        lesson:
          $ref: "#/components/schemas/LessonSummary"
        course_slug:
          type: string
        course_name:
          type: string
        status:
          type: string
          enum: [in_progress, completed]
        updated_at:
          type: string
          format: date-time
    ContinueLearningResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContinueLearningItem"
    Bookmark:
      type: object
      properties:
        lesson_id:
          type: string
          format: uuid
        lesson_title:
          type: string
        course_slug:
          type: string
        course_name:
          type: string
        created_at:
          type: string
          format: date-time
    BookmarkList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Bookmark"
    QuizAttemptRequest:
      type: object
      required: [score, total]
      properties:
        score:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 1
    QuizAttemptResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        score:
          type: integer
        total:
          type: integer
        passed:
          type: boolean
        attempt_number:
          type: integer
    ProfileStats:
      type: object
      properties:
        courses_completed:
          type: integer
        lessons_completed:
          type: integer
        lessons_in_progress:
          type: integer
        bookmarks_count:
          type: integer
        quizzes_taken:
          type: integer
        avg_quiz_score:
          type: [number, "null"]

    # -------------------------------------------------------------- billing
...
    LessonInput:
      type: object
      required: [node_id, title, slug]
      properties:
        node_id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        description:
          type: string
        explanation:
          type: string
        objectives:
          type: array
          items:
            type: string
        examples:
          type: array
          items:
            type: string
        exercises:
          type: array
          items:
            type: string
        quizzes:
          type: array
          items:
            $ref: "#/components/schemas/Quiz"
        level:
          type: [string, "null"]
          enum: [beginner, intermediate, advanced]
        status:
          type: string
          enum: [draft, published]
          default: draft
        sort_order:
          type: integer
          default: 0
    AdminLesson:
      type: object
      properties:
        id:
          type: string
          format: uuid
        node_id:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
        description:
          type: [string, "null"]
        explanation:
          type: [string, "null"]
        objectives:
          type: array
          items:
            type: string
        examples:
          type: array
          items:
            type: string
        exercises:
          type: array
          items:
            type: string
        quizzes:
          type: array
          items:
            $ref: "#/components/schemas/Quiz"
        level:
          type: [string, "null"]
          enum: [beginner, intermediate, advanced]
        status:
          type: string
          enum: [draft, published]
        sort_order:
          type: integer
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        videos:
          type: array
          items:
            $ref: "#/components/schemas/VideoCandidate"
    VideoInput:
      type: object
      properties:
        youtube_url:
          type: string
          description: Full YouTube URL; the ID is extracted and validated server-side.
        youtube_video_id:
          type: string
          pattern: "^[A-Za-z0-9_-]{11}$"
        title:
          type: string
        channel:
          type: string
        is_primary:
          type: boolean
        curator_status:
          type: string
          enum: [pending, approved, flagged, unavailable]
        date_reviewed:
          type: [string, "null"]
          format: date
        notes:
          type: string
    VideoCandidate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lesson_id:
          type: string
          format: uuid
        youtube_video_id:
          type: string
        title:
          type: string
        channel:
          type: [string, "null"]
        is_primary:
          type: boolean
        curator_status:
          type: string
          enum: [pending, approved, flagged, unavailable]
        date_reviewed:
          type: [string, "null"]
          format: date
        notes:
          type: [string, "null"]
        added_by:
          type: string
          format: uuid
    ReviewVideoItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lesson_id:
          type: string
          format: uuid
        lesson_title:
          type: string
        youtube_video_id:
          type: string
        title:
          type: string
        channel:
          type: [string, "null"]
        curator_status:
          type: string
          enum: [flagged, unavailable]
        notes:
          type: [string, "null"]
        last_checked_at:
          type: [string, "null"]
          format: date-time
        is_available:
          type: [boolean, "null"]
...